Extract components into seperate files
